move emotion-registry.tsx
