+++ apps/web/src/components/ui/emotion-registry.tsx
